+++ src/picasapy/app/i18n/picasapy_hu.ts
@@ -5745,6 +5745,11 @@ Egy gyűjtemény megnyitásához kattintson duplán a nevére, vagy kattintson a
     </message>
     <message>
         <location filename="../qml/PicasaPy/EditorCropPanel.qml"/>
+        <source>Discards any applied cropping</source>
+        <translation>Az összes alkalmazott vágás elvetése</translation>
+    </message>
+    <message>
+        <location filename="../qml/PicasaPy/EditorCropPanel.qml"/>
         <source>Landscape</source>
         <translation>Fekvő</translation>
     </message>